Moving files between core and helpers
