+++ iris/api/core/location.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from iris.api.helpers.core_helper import INVALID_NUMERIC_INPUT
+from util.core_helper import INVALID_NUMERIC_INPUT
 
 
 class Location(object):
