+++ iris/api/core/pattern.py
@@ -8,7 +8,7 @@ import os
 import cv2
 import numpy as np
 
-from iris.api.helpers.core_helper import get_module_dir, get_images_path
+from util.core_helper import get_module_dir, get_images_path
 from location import Location
 
 try:
