+++ iris/api/core/platform.py
@@ -4,7 +4,7 @@
 
 import pyautogui
 
-from iris.api.helpers.core_helper import get_os
+from util.core_helper import get_os
 
 
 class Platform(object):
