+++ iris/api/core/util/image_remove_noise.py
