+++ iris/api/core/util/image_search.py
