+++ iris/api/core/util/save_debug_image.py
