+++ iris/test_case.py
@@ -2,6 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
+from api.core.local_web import *
+from api.core.profile import *
 from api.helpers.general import *
 from asserts import *
 from configuration.config_parser import *
